Add blog post parameter to ui-routing
The blog posts are now referred to by /blog/<blogPostId> where blogPostId
is an integer.  This is handled as a parameter so that I do not need to
hard code the route for every blog post.

Additionally, the previous and next uttons for the blog posts now work,
however there may be an easier way to accomplish this functionality
through passing parameters to ui-sref?