Add text overflow handling to class divs
Previously, there were two issues.  If too many words were entered into
the course name (or identically the section number) field, the class div
would grow taller (and take up more 'time' on the schedule) than it was
intended to.  Additionally, if a single very long word was entered, the
div would grow woder than it was intended to and leak into the next days
column.

This issue has been addressed by adding overflow SASS to scheduler.scss
and also specifying the height of the inner span of the div to be 5px less
(padding) than the height of the div itself.