Re-add whitespace to html templates and fix tests
Previously, the whitespace had been removed from several html templates in
tableMaker.coffee because although it made the template easier to read, it
was causing my test to fail.  This however was clearly not the correct
solution, so the whitespace has been replaced and the test modified to
ignore it.

Additionally, a problem was fixed where tableMaker.makeHtml() was
returning an array and scheduler was expecting it to return a string of
pure html.  This was causing commas to appear on the page because the
array was treated like html.  This has been fixed by joining the contents
of the array before returning it.