Reorganize my blog
This involved creating "selection-item" bubbles for the user to click on
to navigate to different sections.  The first page gives the users a list
of categories, which currently consists of "recipes" and "ios".

After selecting a category they go to a similar page where they can select
an item from that category.

Recipes are different than other categories because they have a unique
state structure.