Integrate bootstrap grid system - support mobile
This is a very large commit.  Many things took place inside of this commit
but all with the same goal, actually taking advantage of the bootstrap
that I've been importing into this project since the beginning but never
really using.

Everything is now nicely laid out using bootstrap's fluid-container.  The
nav bar remains on the left while the main page content is centered in the
ramining space.

On smaller screens the nav bar dissappears and is accessable via an
awesome menu button in the upper left corner.

MY WEBSITE IS NOW COMPLETELY MOBILE FRIENDLY, YAYYY!