Add materialize time picker to scheduler
When a course, section, or time is cloned to add another one to scheduler,
the input ids and the label fors are now appropriately updated.

What forced me to actually make this work is that with the materialize
check boxes if the fors weren't right, when you clicked on one checkbox it
should check a completely different one, the first id that matched the
for that you clicked.

Fixing time picker clone problems was done by re-calling the add-crate
listener method at the end of
that method so that the listeners are added again with every clone.  This
might be inefficient?  But it's the only thing I could think of that
works.

Everything works except active state issues.  And the errors aren't
propogating properly to the end user