Redo photoswipe
The previous implementation of photoswipe had been my meager attempt to
copy the photoswipe demo page.  There was a lot of code in a lot of files
that I didn't understand.

The new implementation uses a bower install photoswipe with an
initialization that I wrote using jquery, which is much simpler than what
was happening before.