Fix active issue with schedule timepickers
This was done by adding an afterdone parameter to the options of the
timepicker.

There is still a lesser issue where now all the labels on the page are
given the active label and pop up when only the one should.  But it's
functional and does not make it mor difficult for the user like not having
the active class did by making the label intersect the input contents.